Update ddc_about.md
Added note about the change from TGICA to TG-Data using information from:
https://www.ipcc.ch/task-group-on-data-and-scenario-support-for-impact-and-climate-analysis-tgica/

Added description of the role of TG-Data and the list of TG-Data members using information from:
https://www.ipcc.ch/2019/04/16/task-group-data-members/

I made an educated guess about who the ex-officio members of TG-Data are.  This should be checked and amended before the page goes live.